Вынесены лишние функции из view -> service, добавлена документация для двух функций
